adding UI

+++ main.js
@@ -1,17 +1,30 @@
 // Modules to control application life and create native browser window
-const {app, BrowserWindow, ipcMain, dialog } = require('electron');
+const {app, BrowserWindow, ipcMain, dialog, screen } = require('electron');
 //require('electron-reload')(__dirname);
 let {PythonShell} = require('python-shell');
 const path = require('path');
 const fs = require('fs');
+const sharp = require('sharp');
+const sizeOf = require('image-size');
 
-let mainWindow
+let mainWindow, w, h, dest_image
 
 function createWindow () {
+  // getting system width and height and making the display window full screen
+  let {width, height} = screen.getPrimaryDisplay().workAreaSize;
+
+  if(width < 100){
+    width = 800
+  }
+  if(height < 100){
+    height = 600
+  }
+  w = width
+  h = height
   // Create the browser window.
   mainWindow = new BrowserWindow({
-    width: 800,
-    height: 600,
+    width: width,
+    height: height,
     webPreferences: {
       preload: path.join(__dirname, 'preload.js'),
       contextIsolation: true,
@@ -21,7 +34,7 @@ function createWindow () {
   })
 
   // and load the index.html of the app.
-  mainWindow.loadFile('index.html')
+  mainWindow.loadFile('./pages/index.html')
 
   // Open the DevTools.
   mainWindow.webContents.openDevTools()
@@ -54,18 +67,21 @@ ipcMain.on("toMain", (event, args) => {
     console.log(`From renderer: `+ args[0] + " " + args[1])
     x = args[0]
     y = args[1]
+    temp_dir = path.join(app.getPath('temp'))
+
+    let pypath = path.join(__dirname,'venv','bin','python3.7')
 
     let options = {
       mode: 'text',
       pythonOptions: ['-u'],
-      args: [x, y]
+      pythonPath: pypath,
+      args: [x, y, temp_dir]
     };
     let ret_val;
     let pyshell = new PythonShell(path.join(__dirname,'image_processing','flood-fill.py'), options)
     
     pyshell.on('message', function(message) {
       ret_val = message
-      //console.log("The message: "+ message);
     })
     
     pyshell.end(function (err) {
@@ -74,47 +90,56 @@ ipcMain.on("toMain", (event, args) => {
       };
       console.log("The message: "+ ret_val);
       console.log('finished');
-      /*
-      copyFile(ret_val, dest_image);
-      deleteFile(ret_val);*/
+      // Send result back to renderer process
       mainWindow.webContents.send("fromMain", ret_val);
     });
-    
-
-    // Send result back to renderer process
-    //mainWindow.webContents.send("fromMain", x);
 });
 
 ipcMain.on("chooseFile", (event, arg) => {
-  dest_image = path.join(__dirname, 'image_processing', 'temp_images', 'temp.jpg')
-  /*try {
-    fs.unlinkSync(dest_image)
-    //file removed
-  } catch(err) {
-    console.error(err)
-  }*/
+  //dest_image = path.join(__dirname, 'image_processing', 'temp_images', 'temp.jpg')
+  dest_image = path.join(app.getPath('temp'),'temp.jpg')
   deleteFile(dest_image);
   const result = dialog.showOpenDialog({
     properties: ["openFile"],
-    filters: [{ name: "Images", extensions: ["png","jpg","jpeg"] }]
+    filters: [{ name: "Images", extensions: ["png","jpg","jpeg","gif","png"] }]
   });
   result.then(({canceled, filePaths, bookmarks}) => {
     //Test cancelling
     if(!canceled){
-      //console.log(filePaths[0])
       src_image = filePaths[0]
-      copyFile(src_image, dest_image)
-      mainWindow.webContents.send("chosenFile", dest_image);
-    }
+      //copyFile(src_image, dest_image)
+      if(!checkImageSize(src_image)){
+        mainWindow.webContents.send("invalidFile", "Image is too small, please select a better quality image")
+        return;
+      }
+      sharp(src_image)
+        .resize(800, 600, {
+        kernel: sharp.kernel.nearest,
+        fit: 'contain'
+      })
+      //.toFile(path.join('image_processing', 'temp_images', 'temp.jpg'))
+      .toFile(dest_image)
+      .then(() => {
+        mainWindow.webContents.send("chosenFile", dest_image)
+      });
+      }
   }).catch(err => {
     console.log(err)
   });
 });
 
+//sending image url to renderer
+ipcMain.on("getImage", (event, args) => {
+  mainWindow.webContents.send("returnImage", dest_image);
+});
+
+
 function deleteFile(path_to_file){
   try {
-    fs.unlinkSync(path_to_file)
-    //file removed
+    if(fs.existsSync(path_to_file)){
+      fs.unlinkSync(path_to_file)
+      //file removed
+    }
   } catch(err) {
     console.error(err)
   }
@@ -122,10 +147,20 @@ function deleteFile(path_to_file){
 
 function copyFile(src, dest){
   try {
-    fs.copyFileSync(src, dest)
-    console.log(src+' was copied to '+dest);
+    if(fs.existsSync(src)){
+      fs.copyFileSync(src, dest)
+      console.log(src+' was copied to '+dest);
+    }
   } catch(err) {
     console.error(err)
   }
+}
+
+function checkImageSize(src) { 
+  var dimensions = sizeOf(src);
+  if(dimensions.width < 100 || dimensions.height < 100){
+    return false;
+  }
+  return true;
 }
